[Chore]: Int64 -> Int 타입 변경

+++ Staccato-iOS/Staccato-iOS/Util/Model/Comment/Comment.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 struct Comment: Codable {
-    let commentId: Int64            // 댓글 ID
-    let memberId: Int64             // 회원 ID
+    let commentId: Int            // 댓글 ID
+    let memberId: Int             // 회원 ID
     let nickname: String            // 회원 닉네임
     let memberImageUrl: String?     // 회원 프로필 이미지 URL (옵션)
     let content: String             // 댓글 내용
