+++ Staccato-iOS/Staccato-iOS/Util/Model/Member/Member.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct Member: Codable {
-    let memberId: Int64        // 회원 ID
+    let memberId: Int        // 회원 ID
     let nickname: String       // 회원 닉네임
     let memberImage: String?   // 회원 프로필 이미지 URL (옵션)
 }
