+++ Staccato-iOS/Staccato-iOS/Util/Model/Moment/MomentLocation.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct MomentLocation: Codable {
-    let momentId: Int64      // Moment ID
+    let momentId: Int      // Moment ID
     let latitude: Double     // 위도
     let longitude: Double    // 경도
 }
